requirements.txt
huggingface-hub>=0.20.0
python-libtorrent>=2.0.0
requests>=2.28.0
tqdm>=4.65.0
